Split filters into separate components
